Digital-GetService-main/requirements.txt
Flask==3.1.0
Werkzeug==3.1.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.35
python-dotenv==1.0.1
Flask-Limiter==3.5.0
Flask-Sock==0.7.0
Flask-Session==0.8.0
requests==2.32.3
redis==5.0.7
gunicorn==22.0.0
psycopg2-binary>=2.9.9
pymysql>=1.0.3
waitress==3.0.0
